refactor(cli): split cyclonedx model importer facade
